rename : js 파일 확장자 수정 및 images 폴더명 변경
