+++ src/assets/avatar_default.png
