+++ src/styles/Cards.css
@@ -72,7 +72,7 @@
   right: 2rem;
   width: 2.1rem;
   height: 1.7rem;
-  background: url('../images/icon_kebab.png') no-repeat center/ 2.1rem 1.7rem;
+  background: url('../assets/icon_kebab.png') no-repeat center/ 2.1rem 1.7rem;
   cursor: pointer;
 }
 
@@ -82,12 +82,12 @@
   right: 1.5rem;
   width: 3.4rem;
   height: 3.4rem;
-  background: url('../images/icon_bookmark.png') no-repeat center/ 3.4rem 3.4rem;
+  background: url('../assets/icon_bookmark.png') no-repeat center/ 3.4rem 3.4rem;
   cursor: pointer;
 }
 
 .button-bookmark.filled {
-  background: url('../images/icon_bookmark_filled.png') no-repeat center/ 3.4rem 3.4rem;
+  background: url('../assets/icon_bookmark_filled.png') no-repeat center/ 3.4rem 3.4rem;
 }
 
 @media screen and (max-width: 768px) {
